requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
twilio==9.2.3
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.8.2
